+++ configure.ac
@@ -6,7 +6,7 @@ AC_CONFIG_SRCDIR([src/fabric.c])
 AC_CONFIG_AUX_DIR(config)
 AC_CONFIG_MACRO_DIR(config)
 AC_CONFIG_HEADERS(config.h)
-AM_INIT_AUTOMAKE([1.11 dist-bzip2 foreign -Wall -Werror subdir-objects])
+AM_INIT_AUTOMAKE([1.11 dist-bzip2 foreign -Wall -Werror subdir-objects parallel-tests])
 m4_ifdef([AM_SILENT_RULES], [AM_SILENT_RULES([yes])])
 
 AC_ARG_ENABLE([debug],
@@ -24,7 +24,7 @@ AS_IF([test -z "$CFLAGS"],
 # <3), but it is necessary in AM 1.12.x.
 m4_ifdef([AM_PROG_AR], [AM_PROG_AR])
 
-AM_PROG_LIBTOOL
+LT_INIT([dlopen])
 
 AC_ARG_WITH([valgrind],
     AC_HELP_STRING([--with-valgrind],
@@ -57,6 +57,7 @@ AC_CHECK_DECLS([O_CLOEXEC],,[AC_DEFINE([O_CLOEXEC],[0],
 		#include <fcntl.h>
 		#endif
 	]])
+
 AC_CHECK_DECLS([SOCK_CLOEXEC],,[AC_DEFINE([SOCK_CLOEXEC],[0],
 	[Defined to 0 if not provided])],
 	[[
@@ -110,6 +111,28 @@ AC_CACHE_CHECK(whether ld accepts --version-script, ac_cv_version_script,
 
 AM_CONDITIONAL(HAVE_LD_VERSION_SCRIPT, test "$ac_cv_version_script" = "yes")
 
+AC_ARG_ENABLE([sockets],
+	      [AS_HELP_STRING([--enable-sockets],
+			      [Enable sockets provider @<:@default=yes@:>@])
+	      ],
+	      [],
+	      [enable_sockets=yes])
+
+AS_IF([test x"$enable_sockets" = x"dl"], [sockets_dl=yes; enable_sockets=yes])
+
+AS_IF([test x"$enable_sockets" = x"yes"],
+	[AC_DEFINE([HAVE_SOCKETS], [1],
+		[Define if sockets provider is enabled])
+	],
+	[AC_MSG_NOTICE(Sockets provider not enabled)])
+
+AM_CONDITIONAL([HAVE_SOCKETS], [test x"$enable_sockets" = x"yes"])
+
+AS_IF([test x"$sockets_dl" = x"yes"],
+      [AC_DEFINE([HAVE_SOCKETS_DL], [1], [Build sockets provider as module])])
+
+AM_CONDITIONAL([HAVE_SOCKETS_DL], [test x"$sockets_dl" = x"yes"])
+
 AC_ARG_ENABLE([verbs],
 	      [AS_HELP_STRING([--enable-verbs],
 			      [Enable verbs provider @<:@default=auto@:>@])
@@ -118,36 +141,35 @@ AC_ARG_ENABLE([verbs],
 	      [enable_verbs=auto])
 
 AS_CASE([$enable_verbs],
-	[auto], [AC_CHECK_LIB(rdmacm, rsocket,
-			[AC_CHECK_LIB(ibverbs, ibv_open_device,
+	[auto], [verbs_fail=no],
+	[yes],  [verbs_fail=yes],
+	[dl],   [verbs_dl=yes; enable_verbs=yes],
+	[no],   [],
+	[])
+
+AS_IF([test x"$enable_verbs" != x"no"],
+	[AC_CHECK_LIB(ibverbs, ibv_open_device,
+			[AC_CHECK_LIB(rdmacm, rsocket,
 				[enable_verbs=yes],
 				[enable_verbs=no])
 			],
 			[enable_verbs=no])
-		],
-	[yes],  [AC_CHECK_LIB(rdmacm, rsocket,
-			[AC_CHECK_LIB(ibverbs, ibv_open_device, [],
-				[AC_MSG_ERROR(libfabric requires libibverbs)])
-			],
-			[AC_MSG_ERROR(libfabric requires librdmacm 1.0.16 or greater)])
-		],
-	[no],   [],
-	[])
-
-AS_IF([test x"$enable_verbs" = x"yes"],
-	[AC_DEFINE([HAVE_VERBS], [1],
-		[Define if verbs provider is enabled])
 	],
 	[AC_MSG_NOTICE(Verbs provider not enabled)])
 
-if test x"$enable_verbs" = x"yes"; then
-	AC_CHECK_LIB(rdmacm, rsocket, [],
-		[AC_MSG_ERROR(libfabric requires librdmacm 1.0.16 or greater)])
-	AC_CHECK_LIB(ibverbs, ibv_open_device, [],
-		[AC_MSG_ERROR(libfabric requires libibverbs)])
-fi
+AS_IF([test x"$enable_verbs $verbs_fail" = x"no yes"],
+	[AC_MSG_ERROR(libfabric requires libibverbs, librdmacm 1.0.16 or greater)])
+
+AS_IF([test x"$enable_verbs" = x"yes"],
+	[AC_DEFINE([HAVE_VERBS], [1], [Define if verbs provider is enabled])
+         LIBS=" -libverbs -lrdmacm $LIBS"
+	])
+
+AS_IF([test x"$verbs_dl" = x"yes"],
+	[AC_DEFINE([HAVE_VERBS_DL], [1], [Define if verbs should be built as module])])
 
 AM_CONDITIONAL([HAVE_VERBS], [test x"$enable_verbs" = x"yes"])
+AM_CONDITIONAL([HAVE_VERBS_DL], [test x"$verbs_dl" = x"yes"])
 
 AC_ARG_ENABLE([psm],
 	      [AS_HELP_STRING([--enable-psm],
@@ -198,8 +220,14 @@ AS_CASE([$enable_psm],
 			],
 			[AC_MSG_ERROR([psm_init() not found. Provide the correct path to PSM --with-psm-lib])])
 		],
+	[dl],	[enable_psm=yes; psm_dl=yes],
 	[no],	[],
 	[])
+
+
+AS_IF([test x"$psm_dl" = x"yes"],
+	[AC_DEFINE([HAVE_PSM_DL], [1], [Define if PSM will be built as module])])
+
 
 AS_IF([test x"$enable_psm" = x"yes"],
 	[AC_DEFINE([HAVE_PSM], [1], [Define if PSM is enabled])
@@ -208,6 +236,7 @@ AS_IF([test x"$enable_psm" = x"yes"],
 	[AC_MSG_NOTICE(PSM not enabled)])
 
 AM_CONDITIONAL([HAVE_PSM], [test x"$enable_psm" = x"yes"])
+AM_CONDITIONAL([HAVE_PSM_DL], [test x"$psm_dl" = x"yes"])
 
 AC_ARG_ENABLE([direct],
 	[AS_HELP_STRING([--enable-direct=@<:@provider@:>@],
