+++ src/common.c
@@ -1,0 +1,144 @@
+/*
+ * Copyright (c) 2004, 2005 Topspin Communications.  All rights reserved.
+ * Copyright (c) 2006 Cisco Systems, Inc.  All rights reserved.
+ * Copyright (c) 2013 Intel Corp., Inc.  All rights reserved.
+ *
+ * This software is available to you under a choice of one of two
+ * licenses.  You may choose to be licensed under the terms of the GNU
+ * General Public License (GPL) Version 2, available from the file
+ * COPYING in the main directory of this source tree, or the
+ * OpenIB.org BSD license below:
+ *
+ *     Redistribution and use in source and binary forms, with or
+ *     without modification, are permitted provided that the following
+ *     conditions are met:
+ *
+ *      - Redistributions of source code must retain the above
+ *        copyright notice, this list of conditions and the following
+ *        disclaimer.
+ *
+ *      - Redistributions in binary form must reproduce the above
+ *        copyright notice, this list of conditions and the following
+ *        disclaimer in the documentation and/or other materials
+ *        provided with the distribution.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#if HAVE_CONFIG_H
+#  include <config.h>
+#endif /* HAVE_CONFIG_H */
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include <fcntl.h>
+#include <unistd.h>
+#include <poll.h>
+#include <pthread.h>
+
+#include <rdma/fi_errno.h>
+#include "fi.h"
+
+int fi_wait_cond(pthread_cond_t *cond, pthread_mutex_t *mut, int timeout)
+{
+	struct timespec ts;
+
+	if (timeout < 0)
+		return pthread_cond_wait(cond, mut);
+
+	clock_gettime(CLOCK_REALTIME, &ts);
+	ts.tv_sec += timeout / 1000;
+	ts.tv_nsec += (timeout % 1000) * 1000000;
+	return pthread_cond_timedwait(cond, mut, &ts);
+}
+
+int fi_read_file(const char *dir, const char *file, char *buf, size_t size)
+{
+	char *path;
+	int fd, len;
+
+	if (asprintf(&path, "%s/%s", dir, file) < 0)
+		return -1;
+
+	fd = open(path, O_RDONLY);
+	if (fd < 0) {
+		free(path);
+		return -1;
+	}
+
+	len = read(fd, buf, size);
+	close(fd);
+	free(path);
+
+	if (len > 0 && buf[len - 1] == '\n')
+		buf[--len] = '\0';
+
+	return len;
+}
+
+int fi_poll_fd(int fd, int timeout)
+{
+	struct pollfd fds;
+
+	fds.fd = fd;
+	fds.events = POLLIN;
+	return poll(&fds, 1, timeout) < 0 ? -errno : 0;
+}
+
+struct fi_info *__fi_allocinfo(void)
+{
+	struct fi_info *info;
+
+	info = calloc(1, sizeof(*info));
+	if (!info)
+		return NULL;
+
+	info->ep_attr = calloc(1, sizeof(*info->ep_attr));
+	info->domain_attr = calloc(1, sizeof(*info->domain_attr));
+	info->fabric_attr = calloc(1, sizeof(*info->fabric_attr));
+	if (!info->ep_attr || !info->domain_attr || !info->fabric_attr)
+		goto err;
+
+	return info;
+err:
+	__fi_freeinfo(info);
+	return NULL;
+}
+
+void __fi_freeinfo(struct fi_info *info)
+{
+	free(info->src_addr);
+	free(info->dest_addr);
+	free(info->ep_attr);
+	if (info->domain_attr) {
+		free(info->domain_attr->name);
+		free(info->domain_attr);
+	}
+	if (info->fabric_attr) {
+		free(info->fabric_attr->name);
+		free(info->fabric_attr->prov_name);
+		free(info->fabric_attr);
+	}
+	free(info->data);
+	free(info);
+}
+
+uint64_t fi_tag_bits(uint64_t mem_tag_format)
+{
+	return UINT64_MAX >> (ffsll(htonll(mem_tag_format)) -1);
+}
+
+uint64_t fi_tag_format(uint64_t tag_bits)
+{
+	return FI_TAG_GENERIC >> (ffsll(htonll(tag_bits)) - 1);
+}
+
